rotate_3d_model/requirements.txt
boto3~=1.20.34
Pillow~=9.0.0
pip~=21.3.1
numpy~=1.22.0
numpy-stl~=1.6.2
requests~=2.27.1
botocore~=1.23.34
s3transfer~=0.5.0
jmespath~=0.10.0
setuptools~=41.2.0
yarg~=0.1.9
certifi~=2021.10.8
docopt~=0.6.2
pipreqs~=0.4.11
six~=1.16.0
python-dateutil~=2.8.2
urllib3~=1.26.8
matplotlib~=3.5.1
idna~=3.3
fonttools~=4.28.5
pyparsing~=3.0.6
cycler~=0.11.0
packaging~=21.3
kiwisolver~=1.3.2